fix(devctr): copy only poetry files during build
During the container build, all files in the devctr folder are copied
into /tmp/poetry, including the Dockerfile itself. When rebuilding a
Dockerfile, Docker is smart and only rebuilds the parts that changes,
but since the Dockerfile itself is copied over, the build starts from
scratch every time and doesn't reuse intermediate images from earlier
builds. This is annoying when adding new things to the dev container as
every attempt will incur in a long rebuild process.
By copying only the files relevant to poetry, we allow Docker to only
rebuild the commands that changed.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>