feat(metrics): store dimensions alongside metrics in store_data
In d6e5521b we added a metrics.json to dump all emitted metrics, but we
don't add the dimensions to easily parse and know to know to what they
refer.
This patch changes the format of the metrics.json to add two keys:
metrics, contanining the old dict of metrics, and dimensions,
contiaining the dict of dimensions.
We're not storing all other properties as they are the same across all
tests on the same executor and they can be fetched from test_result.json
already, under "user_properties".

Signed-off-by: Riccardo Mancini <mancio@amazon.com>