test(network_latency): drop check on number of pings returned
test_network_latency uses ping to send 15 rounds of 30 pings from the
guest to the host and collects the latencies printed. For this, instead
of just parsing the ping output for however many lines of latency data
there we, implicitly assert that we get exactly 30. However, sometimes
this fails. Nikita and Egor investigated and found that no packages are
lost, sometimes ping just seems to not print all 30 lines for some
reason (presumably a bug in ping). So just drop this assertion to stop
spurious test failures from happening. This is a performance test, not a
functional test anyway.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>