test(balloon): ensure we always wait long enough for stats to update
Always wait twice the stats polling interval when trying to a comparison
between "old" and "new" balloon stats. This is to ensure that the "new"
stats are definitely new, and we didn't hit some unlikely race condition
where we request stats right as the device is still in the process of
updating them.

We were seeing intermittent issues such as

test_balloon.py:531: in test_balloon_snapshot
    assert stats_after_snap["available_memory"]
           > latest_stats["available_memory"]
E   assert 186281984 > 186281984

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>