test: Add test for Go SDK credential provider with custom endpoint
AWS SDK for Go sets "Accept: application" in a request to retrieve AWS
credentials. If imds_compat is true, it should work. If false, it should
NOT work, because MMDS responds a string of a JSON object containing the
credentials (i.e. wrapped with doulequotes) with "Content-Type:
application/json" but AWS SDK for Go expects only the inner JSON object.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>