fix: return NO_VECTOR when reading MSI vector for invalid queue
Fix a bug in the common VirtIO configuration for PCI transport where we
would use `queue_select` to read the queue's MSI vector without
validating it matches a valid queue. This could lead in panics when
accessing the `msix_queue` array. The spec states that in such cases we
should return `NO_VECTOR` (0xffff), so do that.

Signed-off-by: Babis Chalios <bchalios@amazon.es>