fix(ci): Dont run functional tests if changing patch series
Currently, we often get stuck with the problem where something in the
host kernel breaks that causes functional tests to fail, but we cannot
update the patch series from which the host kernel gets built, because
functional tests are failing. Break this cyclic dependency by simply not
running functional tests when updating only the patch series (as they
dont test the updated kernel anyway.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>