pci: remove unused type from PciDevice::allocate_bars
Cloud Hypervisor code is passing a resource type argument in the logic
that allocates BARs for devices. `Resource` is an Enum where one of its
variants is `PciBar`. Not sure what Cloud Hypervisor uses this for, but
it seems redundant since this method is specifically used to allocate
BAR memory for devices.

We definitely don't use it, so remove it.

Signed-off-by: Babis Chalios <bchalios@amazon.es>