feat(virtio/interrupt): add trigger_queues method
This method is used to notify the guest about queue events in a way
that's most performant with the underlying interrupt implementation.
As in IrqTrigger there is no distinction between different queues, it's
best to send just one interrupt notifiying that "some queues" have a
pending event.
Conversely, in VirtioInterruptMsix, we need to trigger a MSI for each
distinct queue.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>
Signed-off-by: Babis Chalios <bchalios@amazon.es>