fix(vsock): restore previous performance with PCI disabled
We noticed that, compared to main, the vsock device is up to 30% slower
when PCI is disabled. This is due to the refactor in 087e185
("fix(vsock): pass correct index when triggering interrupts"), as we're
now sending interrupts as soon as we detect the event, rather than
deferring them and sending only one interrupt at the end.
While with MSI we need to send multiple interrupts, so there is little
difference (still, tests show up to 5% improvement with this change),
with legacy IRQ there's only one interrupt line so we end up sending
multiple back to back interrupts rather than a single one.
This patch reverts to the previous behaviour and uses the newly
introduced `trigger_queues` method to deduplicate interrupts in the case
of IrqTrigger.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>
Signed-off-by: Babis Chalios <bchalios@amazon.es>