test: enable serial console in integration tests
If SSH to guest is failing, we don't really get much useful in terms of
logs in our ci artifacts. Enable serial console so that we get guest
dmesg always.

Since serial console is enabled by default now, this means we can stop
explicitly passing modified bootargs in a bunch of tests now.

Disable it for the tests that interact with the serial console via
stdin/stdout (here we are not daemonizing, so we are capturing serial
output anyway).

Disable pylint's "too-many-statements" lint, because it started firing
in microvm.spawn()

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>