pci: add unit test for configuring MSI-X capability
Make sure that configuring the MSI-X capability in PCI configuration
space works properly (configuration space is initialized as expected).
Also, make sure that the implementation respects the read/write
properties of the respective bits.

Finally, fix logic in the code that adds capabilities to take into
account properly the size of capabilities.

Signed-off-by: Babis Chalios <bchalios@amazon.es>