test: hack: tmp: disable serial for A/B tests
When A/B-testing, the "A" revision firecracker binary does not know
about the new --serial-out-path CLI argument. Thus, disable the serial
console for all A/B-tests. It gets quite ugly to do for the
vulnerability tests, because the .spawn() call is burried deep in the
fixtures.

This commit should be reverted after merging the PR

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>