refactor: use setup_serial_device() on aarch64 restoration path
Have teh aarch64 serial device restoration path use the common serial
device initialization code in setup_serial_device(). The only functional
change here is that now we do set_stdou_nonblocking() on the restore
path, although I would wager that not doing so in the past was actually
an oversight, because we do it for all booted VMs, and for restored VMs
on x86.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>