fix(virtio): avoid panic on device activation failure
According to the VirtIO spec section 2.1.2, when we enter a failed state
we need to to set `DEVICE_NEEDS_RESET` status field and if `DRIVER_OK`
is set we need to must send a device configuration change interrupt to
the driver.

In MMIO code we were trying to follow this logic, but we were trying to
get the interrupt object from the device, which fails (and panics)
because interrupts are only set in the device upon successful
activation. In the PCI transport, instead, we were not doing this at
all.

The transport layers hold a reference to the interrupts at all times.
So, add the missing logic to the PCI transport and use the transport
layer reference in both transports to avoid panics.

Signed-off-by: Babis Chalios <bchalios@amazon.es>