chore(bindgen): upgrade to linux 6.12 headers and use single kernel tree
We were using bindgen from 3 different sources:
 - installed headers in the system
 - linux-5.10.y branch from amazonlinux repo
 - linux-next branch from torvalds repo

This is overcomplicated and unnecessary as Linux doesn't break the UAPI.

This patch updates it to use a single stable branch (6.12), from which
the user headers are generated using `make headers_install` and
regenerates all the headers, fixing some minor incompatibilities in the
syntax.

While I was at it I also improved the generation for C enums to use
constified-enum-module style which is less verbose.

I had to fix an issue with the io_uring auto-generated structs not
implementing Copy due to FAM introduced in the latest kernels. I have
put a bindgen-patch to drop the FAM and derive Copy and Clone where
required to implement ByteValued. If we ever need to use those fields,
we'll need to find a different solution.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>