fix(virtio-pci): check guest values for MSI-X vector
We should check that the MSI-X vector that a guest assigns to a VirtIO
queue or configuration is a valid one. If it is a value bigger than the
available vectors allocated for the device, it should not update the
vector field and mark the corresponding vector to the NO_VECTOR value.

Also, add checks in the MSI implementation for VirtioInterrupt trait,
for making sure that we don't try to interact with out-of-range
interrupt vectors.

Signed-off-by: Babis Chalios <bchalios@amazon.es>