refactor: cleanup vmm::snapshot module
Implement the suggestions from #4523 to give a cleaner API for the
snapshot module. Conceptually, the change is that we now store the thing
we are snapshotting inside of the `Snapshot` struct.
Implementation-wise, this allows us to deserialize the snapshot data in
a single pass (instead of having to read it twice, once for
deserialization and once for CRC validation), which means we do not have
to first query the snapshot file for the size of the snapshot.

Closes #4691, #4523, #5195

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>