chore(clippy): enable warn for or_fun_call
"or_fun_call" warns us when a function is called in side a "*_or(...)"
function. In this case the value is computed (which may be expensive),
but is only used if the Result/Option is an Error/None.
In these cases, using the "*_or_else" variant is the best thing to do.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>