dependabot: drop open-pull-requests-limit
Having it set to 100 is not very useful (the number is from before we
were using grouped updates). With the way dependabot is configured, we
can at most get 2 PRs at the same time (as dependabot will only open at
most one PR per group, and we have two groups). We could set it to 2,
but that can cause problems, see
https://github.com/rust-vmm/kvm/pull/336.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>