fix(pci): fix handling of VirtIO PCI capabilities
The VirtioPciCap type had the wrong structure. This wasn't a problem,
because the wrong field was actually operating as part of the padding
of which it was taking away the space.

Also, fix the initialization of the VirtioPciCfgCap so that it reports
the correct length for the capability.

(cherry picked from commit 8ee3af1fa6d5153da317051e9045ddbcdd3324ed)

Signed-off-by: Babis Chalios <bchalios@amazon.es>