fix(pci): handling of PCI ISR capability
We always enable MSI-X interrupts for VirtIO devices when using the PCI
transport. Currently, we don't offer a fallback if the driver doesn't
want to use MSI-X, but VirtIO drivers we are working with always support
that. So this should not be a problem for our use-cases in the kernels
we support at the moment.

On the other hand, despite not supporting legacy interrupts, we still
did some maintenance for the ISR status byte which is not used by MSI-X
interrupts at all. Drop all this handling, ensure that reads always
return 0 and writes have no effect in the state of the device. Also, add
warning messages for the cases where the guest tries to access the ISR
parts of the BAR.

Signed-off-by: Babis Chalios <bchalios@amazon.es>