pci: document potential contention on PCI bus lock
We have a single lock for all devices on the PCI bus that serializes
reads and writes on the devices' PCI configuration space and BARs.

This should not be a problem at the moment. It should be out of any hot
path and only up until we are setting up devices. However, add a comment
that mentions the existence of the contention so that we keep it in mind
for the futuer (and maybe perform some profiling).

Signed-off-by: Babis Chalios <bchalios@amazon.es>