fix(getting-started): increase ext4 size and verify artifacts
We increased the ext4 size from 400M to 500M in our devtool scripts,
but not in the getting started. Let's bump it to 1G to avoid issues like
this in the future.

In the longer term, we may look into simplifying the getting started
guide, but for now it's quicker to just fix the issue.

Fixes: #5349
Signed-off-by: Riccardo Mancini <mancio@amazon.com>