refactor(pci): assume we only have a single 64-bit BAR
In our case, only VirtIO PCI devices use a BAR region. Moreover, this
region has a fixed size and it's always a 64-bit region. This means we
don't handle 32-bit or IO BARs and we don't support a ROM bar. Encode
these assumptions to the logic that adds BAR regions and detects BAR
reprogramming attempts from the guest side so that we simplify the code.

Signed-off-by: Babis Chalios <bchalios@amazon.es>