test(perf/vsock): copy iperf binary only once
Problem: Currently we copy the iperf3-vsock binary into the guest in
spawn_iperf3_client that is called by every client (we have 2).  If by
the time the second binary is being copied that first one is already
executing, we get the "Text file busy" error.

Solution: Move the copy into the test construction to copy the binary
only once.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>