fix(test): correct operator precedence
bitshift shift has lower precedence than division, meaning in
MemoryUsageExceededError.__init__ we end up trying to bitshift a float,
which is nonsense (as the expression in the f-string gets interpreted as
(threshold / 1) << 20, and the result of all divisions in python is a
float). Add some parenthesis.

We cannot just write this as `threshold >> 20`, as we care about the
fractional part.

Fixes: 7d1549fccd1d ("tests: fix MMIO gaps in memory monitor tool")
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>