fix(test): disable memory monitor in uvm_restored
In the uvm_restored fixture, we create a throwaway VM to take a snapshot
of. This VM is completely invisible to the test, so cannot be configured
differently. If the memory monitor triggers in this VM, then it has
nothing to do with the test itself, and it is not recoverable. So just
disable the memory monitor for this VM.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>