fix(test): do not override Firecracker binary
test_redundant_seccomp_rules builds the Firecracker binary using the
nightly toolchain (we need this to be able to run its static analysis).
This replaces the binary used by all subsequent tests and it causes them
to fail since nightly might introduce a system call that is not allowed
by our seccomp rules.

In order to avoid such situations, build the binary used by the test in
a different build directory.

Signed-off-by: Babis Chalios <bchalios@amazon.es>