test: clean-up potentially dead microVMs
We have some tests that stress the microVM in a way that _might_ cause
it to die. We do check whether that's the case, however this is
inherently racy, i.e. the microVM might die on us after the check but
before the fixture logic tries to recycle it. Such a test is
test_balloon.py::test_deflate_oom.

In order to avoid such situations, teach MicroVM.kill() that there might
be cases where the microVM process has already died and let it finish
its cleanup gracefully.

Signed-off-by: Babis Chalios <bchalios@amazon.es>