fix: test_failing_filter intermittent issue
We intermittently get failures in this test while destroying the microVM
object given to us by the fixture. That code checks that the microVM was
not killed due to a signal and in this case, it clearly has (that's the
intention of the test). The test itself is marking the microVM as killed
so the cleanup logic should never reach that check.

The only way that kill() fails in this case is when
test_microvm.mark_killed() (the last command in this test) is never
called, meaning that the test errors out before that. I believe that the
only point in the code that can cause the test to exit earlier is this:

```
try:
    test_microvm.start()
except requests.exceptions.ConnectionError:
    pass
```

We are expecting test_microvm.start() to fail due to seccomp violations,
however we are only catching the ConnectionError exception. If any other
exception is raised the test will fail pre-maturely without setting
test_microvm.mark_killed(). In fact, looking at the backtrace from
intermittent failures it looks like the code is actually raising a
ConnectionResetError.

Change the test to catch any error, rather than just ConnectionError.

Signed-off-by: Babis Chalios <bchalios@amazon.es>