refactor(balloon): move remove_range to GuestMemoryExtension
Move the function to discard guest memory under the GuestMemoryExtension
trait.
This is done so that it could be reused in other parts of the code.

Furthermore, this also changes it to use try_access, which is the
correct way to access a range of guest memory as it deals with the
possibility of it spanning multiple regions.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>