feat(virtio-mem): implement virtio requests
This commit adds block state management and implements the virtio
requests for the virtio-mem device.

Block state is tracked using a BitVec, each bit representing a single
block.

Plug/Unplug requests are validated before being executed to verify the
range is valid (aligned and within range), and that all blocks in range
are unplugged/plugged, as per the virtio spec.

UplugAll is the only request where usable_region_size can be lowered.

This commit is missing the dynamic KVM slot management which will be
added later.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>