refactor(pci): move PCI emulation logic in vmm mod
The implementation of the PCI emulation components is heavily
Firecracker opinionated. Move those inside a module of `vmm` crate and
leave in the `pci` module only type definitions which can easily be
reused across VMMs.

We also take advantage of this refactoring to remove in various places
the usage of `std::io::Error` as error types returned by emulation
logic.

Signed-off-by: Babis Chalios <bchalios@amazon.es>