fix(examples/uffd): unregister range on UFFD Remove event
If the handler receives a UFFD remove event, it currently stores the PFN
and will reply with a zero page whenever it receives a pagefault event
for that page.
This works well with 4k pages, but zeropage is not supported on
hugepages. In order to support hugepages, let's just unregister from
UFFD whenever we get a remove event. By doing so, the handler won't
receive a notification for the removed page, and the VM will get a new
zero page from the kernel.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>