fix(pci): correctly set device status to 0 on reset
Even though currently, no device supports resetting, when the guest
attempts to reset the device we must set the status to 0 as
specified in the 2.4.1 Device Requirements: Device Reset.

This does not change the current status of us not supporting device
resetting. This is the issue we encountered during addition of
virtio-pmem device where the guest kernel now resets devices on VM
shutdown.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>