fix(tests): do not look into full path for error metric
In order to collect all error metrics we're checking for "err" in the
key name, which causes performance tests to fail because we added
metrics called "intERRupts". Change the check to only look at the leaf
of the path, so that when we have a metric such as
`fc_metrics.net_metrics.cfg_fails`, we'll only be looking into
`cfg_fails`.

Suggested-by: Riccardo Mancini <mancio@amazon.co.uk>
Signed-off-by: Babis Chalios <bchalios@amazon.es>