chore: Update fingerprint for CPUID.0x80000021:ECX on AMD / 5.10 kernel
KVM added support for CPUID leaf 0x80000021 [1] and synthesized
TSA_{SQ,L1}_NO bits on ECX [2]. However, the second patch has a bug
where TSA_{SQ,L1}_NO bits are 1st and 2nd bits [3] but exposed as 11th
and 12th bits. This bug happened because it used to be software-defined
word (not hardware-defined). Recently, the kernel added another bit
(X86_FWEATURE_IBPB_EXIT_TO_USER) for VMScape that is not a hardware-
defined bit but a software-defined bit.

We update the fingerprints for now to make the fingerprint test pass,
but will fix the bug. So we'll need to update again once the fix
arrives.

[1]: https://github.com/amazonlinux/linux/commit/6457a8c0a02694375271cb40250527672c64395d
[2]: https://github.com/amazonlinux/linux/commit/6fea1a4e150f1fe3726310558f33a30e0c209153
[3]: https://www.amd.com/content/dam/amd/en/documents/resources/bulletin/technical-guidance-for-mitigating-transient-scheduler-attacks.pdf
[4]: https://github.com/amazonlinux/linux/commit/ac60717f9a8d21c58617d0b34274babf24135835

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>