chore: Update fingerprint for VERW_CLEAR bit pass through
The kernel patch [1] passed through CPUID.0x80000021:EAX[5] (VERW_CLEAR)
to tell guests that the microcode is applied and the memory form of the
VERW instruction can be used to clear the microarchitectural data
structures necessary to mitigate TSA-L1 and TSA-SQ. Thanks to it, we can
drop the exception for vulnerabilities sysfs check inside guest. In
addition to the VERW_CLEAR bit passthrough, the kernel also started
always setting CPUID.0x80000021:EAX[9] (NO_SMM_CTL_MSR) since SMM_CTL
MSR is not available for KVM guests.

[1]: https://github.com/amazonlinux/linux/commit/8d1e0db16431610b5b35737d88595bdd7a08e271
Signed-off-by: Takahiro Itazuri <itazur@amazon.com>