refactor: remove SharedDeviceType
This wrapper was only used in the `update_from_restored_device`
function which was instantly unwrapping the value and
just adding device to the VmResources. Instead of this
round trip just add devices directly since fields are
public.
This also removes the check for the addition of balloon device,
but this check is not needed anyway, since snapshot should not
contain conflicting configs.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>