fix(memory): ensure we are not mmapping beyond EOF on restore from file
vm-memory@0.17.0 removed the bounds check in the memory region creation
as not every fd is seekable. This causes a SIGBUS if for some reason the
snapshot file is smaller than the sum of the regions and the guest tries
to access a page beyond EOF. This issue was caught by the fuzzer.

This change adds the bounds check in the snapshot_file function and unit
tests to ensure this behaviour is maintained.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>