fix(iovec): do not return error if some bytes were written
When reading from a volatile memory into a WriteVolatile we should let
the caller know if we managed to transfer some bytes. Currently, if the
write encounters any error after the first iovec an error is returned,
but some bytes are written. This causes issues in the caller, because it
thinks no bytes were written.

In particular, this caused a bug in the vsock code on 6.17 guest kernels
because these kernels started sending multiple iovecs. When sending 2
iovecs, it could happen that we'd get a WouldBlock error on the second,
but the vsock code thought no bytes were transferre, leading to a
spurious retransmission of the first iovec.

Fixes: #5475

Signed-off-by: Riccardo Mancini <mancio@amazon.com>