vmclock: add support for VMclock device for x86_64
Implement the VMClock device on x86_64 platforms. At the moment, we just
allocate the memory region in the guest address space for exposing the
device. We don't expose any clock from the host and since we don't
support live migration, the device won't do anything at the moment, but
we should still be able to see a `/dev/vmclock` inside the guest.

We do support the `disruption_marker` field which notifies the guest to
adjust clocks due to a time shifting event.

Signed-off-by: Babis Chalios <bchalios@amazon.es>