acpi: simplify the construction of VMGenID device
Decouple creation of VMGenID device with writing the generation ID in
guest memory. This way we can avoid keeping an `Option<VmGenId>` inside
the ACPI device manager. We can always create it and only write the
generation ID in guest memory once we are ready to activate the device
(when we have a functioning guest memory object).

Also, remove a few of the error types and fail in-place instead of
propagating errors all the way up.

On the Aarch64, we need to recreate the expected dtb files because
default microVMs always create entries for the VMGenID device.

Signed-off-by: Babis Chalios <bchalios@amazon.es>