fix(test): skip test_cpu_topology test on Ubuntu
On Aarch64 systems we do not normalize vCPUs, we let KVM configure them.
This means that the CPU cache hierarchy exposed to the guest depends
entirely on what the KVM decides it to be.

We observed that when we're running tests on Ubuntu hosts with a 6.14
kernel the cache hierarchy is different than the one we expect on AL
hosts.

Until we decide how to deal with CPU hierarchy in a controled way (or
not) skip the test on Ubuntu systems. The actual test checks that the
host kernel version is greater or equal to 6.14.

Signed-off-by: Babis Chalios <bchalios@amazon.es>