fix: store bitvec as Vec<bool> to prevent fuzzer crashes
When the fuzzer generates an invalid bitvec state, the serde code
crashes with an assertion error.
To avoid these crashes, just store it as a plain Vec<bool> as we don't
expect these vecs to be too big (for a 10GB hotpluggable area with 2MiB
blocks it would require 5kB for the Vec<bool> compared to 640B with
BitVec).

Signed-off-by: Riccardo Mancini <mancio@amazon.com>