fix(memory): validate memory region state before resuming
When the region state is invalid or corrupted (like when generated by
the fuzzer), it is possible that a DRAM slot is unplugged, leading to
segfaults when accessing guest memory (ie from vmgenid device).
To avoid these crashes, validate the region state and allow the DRAM
region (not hot-pluggable) to only contain one plugged slot.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>