virtio-balloon: Drop huge page restriction
While the traditional balloon device would not be able to reclaim memory
when back by huge pages, it could still technically be used to to
restrict memory usage in the guest.

With the addition of hinting and reporting, they report ranges in bigger
sizes (4mb by default). Because of this, it is possible for the host
reclaim huge pages backing the guest.

Updates the performance tests for the balloon when back by huge pages,
added varients to the size reduction tests to ensure hinting and
reporting can reduce the RSS of the guest.

Move the inflation test to performance to ensure it runs sequentially in
CI otherwise the host can be exhausted of huge pages.

Signed-off-by: Jack Thomson <jackabt@amazon.com>