feat: add numpy, pandas and seaborn to poentry.toml
These libraries are needed for the ab_plot.py.
At the same time update `python` version defined in the
`pyproject.toml`. This is required because for some reason
poetry errors when pooling dependencies saying the python
version is too old. Updating the python version caused
other problems with poetry installation with `pip3 install`.
Because of this, change the way poetry is insatlled by
pooling it directly from a website. Additionally the change
broke somehow `virtualenv`, so just install `python3-venv` manually.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>