fix(gdb): remove wrong ordering and double locking
With the vcpu refactoring gdb support got broken. While we have a test
to verify it builds, we still don't have one to verify it works.
In particular, the gdb thread was created too early, and the vmm was
locked twice leading to a deadlock.

This change fixes both issues and adds a little bit more logging for the
errors in GDB which were helpful to debug the issue.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>