test(swagger): validate API requests against Swagger in integ tests
Validate that all requests we do to the API are conformant to the
swagger, and that all responses are also conformant.
This check is strict, meaning no extra fields are allowed to catch
problems where there is a typo in the swagger.
We only check successful requests as we don't want to fail when we try
to send a bad request on purpose. If the request is successful, then it
means the schema should have been valid.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>