test(gdb): add integ test to verify gdb works on x86
The integ test opens gdb on a subprocess which connects to Firecracker,
sets a breakpoint and continues until boot is complete. If the VM boots
correctly and the breakpoint gets hit, then the test will succeed.

This test currently only works on x86 as GDB requires a vmlinux image
(we use a uImage on ARM).

Signed-off-by: Riccardo Mancini <mancio@amazon.com>